[Implementation][Testing]
Modified implementation of the object and API classes to synchronize the
work that Emma and Jonathan had done.

Commented out methods in object classes that perform operations on the
services because I will try to move that implementation to the APIs.

Performed unit testing on the Customer and Employee Classes.

TO-DO: testing on the API Employee and Customer classes.