[Refactoring][Implementation][Test] Implemented Hashed Passwords.
Added hashing method to User to store the passwords as hashed passwords.
The checkPassword method was needed to compare the hashed passwords.

Created UserInfo class to create custom transfer of data between front
and back end. This prevented passwords from being hashed twice and
allowed specific user data to be sent back to front end (not password).

I updated the EmployeeTest and CustomerTest to reflect these changes and
to properly test the new hash functionality.

I added the validateUser endpoint to the Customer and Employee APIs for
validating a user's log in credentials and tested them in their test
classes.

I, then, updated the front end log in files to use this new endpoint and
make sure it worked.