[Bug Fix]
To fix the issue of the buttons not working, it seems like when some
things were getting logged to the console, "$scope." was not added in
front of some variable names like "ingredient" or "newIngredient". Next,
when making the POST API call to APIInventoryController to add a new
ingredient1, the path was "/api/v1/inventory/" when the path didn't need
the last / character. To reduce redundancy, I defined $scope.ingredient
and $scope.newIngredient in $scope.reset(). Also when adding a new
ingredient, I think the "newName" and "newAmount" variables were
throwing the database and APIInvetoryController off, as the code would
search for "name" and "amount" fields, rather than "newName" and
"newAmount". Therefore, I updated their names. Lastly, I made the form
submission buttons with the input class where their type is "submit"
which'll automatically call the defined form submission function.