[Implementation][Test] Enhanced OrderTest.
Added a test method to test the OrderService getOrder methods. This
required a change of implementation to use findByStatus() instead.

Changed picked_up to a field of Order instead of it's status so that
orders that are complete and/or picked up will all be returned by the
same status.