[Implementation][Test] Made this branch up to date.
Copied over the OrderService and OrderRepository implementations.

Fixed up the Order, OrderStatus, and OrderTest implementations.
- had to modify the equals test method because the id doesn't get
assigned until after orders are saved so it threw an error.

Updated the APIOrderTest to accommodate changes and to continue testing.

Modified APIOrderController implementation:
- created 2 delete mappings, one for customer and one for employee
- had to add /customer/ to the getCustomerOrder path bc it was too
similar to the getOrder path and it complained.
- Made other small modifications to the conditions checked and return
entities.

TO-DO: APIOrderTest not passing because id is null even though it is getting posted and therefore saved.